remove wow legacy test cases
