+++ src/api/host.rs
@@ -1,6 +1,5 @@
-use super::user::{HandshakeProof, StrongProofVerifier};
-use crate::primitives::*;
-use crate::{Result, Srp6Error};
+use crate::prelude::*;
+use crate::Result;
 use serde::{Deserialize, Serialize};
 
 /// this trait provides a higher level api
@@ -205,167 +204,19 @@ impl<const KEY_LENGTH: usize, const SALT_LENGTH: usize> Handshake<KEY_LENGTH, SA
     }
 }
 
-#[cfg(test)]
-pub mod tests {
-    use super::*;
-    use crate::StrongSessionKey;
-    use std::convert::TryInto;
-
-    pub struct Srp6_256Mock;
-    impl Srp6_256Mock {
-        #[allow(non_snake_case)]
-        pub fn I() -> UsernameRef<'static> {
-            "ADMINISTRATOR"
+#[test]
+fn should_panic_when_key_length_does_not_fit_to_modulus() {
+    type Srp = Srp6<10, 10>;
+    let err = Srp::new(
+        Generator::from(3),
+        PrimeModulus::from_hex_str_be("FE27").unwrap(),
+    );
+    assert!(err.is_err());
+    assert_eq!(
+        err.err().unwrap(),
+        Srp6Error::KeyLengthMismatch {
+            expected: Srp::KEY_LEN,
+            given: 2
         }
-
-        pub fn p() -> ClearTextPasswordRef<'static> {
-            "ADMINISTRATOR"
-        }
-
-        pub fn v() -> PasswordVerifier {
-            "3E9D557B7899AC2A8DEC8D0046FB310A42A233BD1DF0244B574AB946A22A4A18"
-                .try_into()
-                .unwrap()
-        }
-        pub fn b() -> PrivateKey {
-            "ACDCB7CB1DE67DB1D5E0A37DAE80068BCCE062AE0EDA0CBEADF560BCDAE6D6B9"
-                .try_into()
-                .unwrap()
-        }
-
-        #[allow(non_snake_case)]
-        pub fn B() -> PublicKey {
-            "35A59FAEEBBE7E45204EE68B16430F9C999392DA17931148B249290FF8F5BBF6"
-                .try_into()
-                .unwrap()
-        }
-
-        #[allow(non_snake_case)]
-        pub fn A() -> PublicKey {
-            "1FFB26380AB82F00FB01A7EBC01C4967E76DC129FF35174E4C6D3C190C36D697"
-                .try_into()
-                .unwrap()
-        }
-
-        pub fn s() -> Salt {
-            "A67FFC191C7EF028F4351E5AE7B65B817448E9F904B4DC7A6572E7F23C8558D5"
-                .try_into()
-                .unwrap()
-        }
-
-        #[allow(non_snake_case)]
-        pub fn S() -> SessionKey {
-            "1EC9B973F4EEDBA8441C0F52FEF4191154B31CAF4D6D30B1D0BABB5CD0385F67"
-                .try_into()
-                .unwrap()
-        }
-
-        #[allow(non_snake_case)]
-        pub fn K() -> StrongSessionKey {
-            "260473D7AEF9043C2E8C65C2CCF479E948D213E3CD00BC869BE64E530D4662A9A4168DDC0A834713"
-                .try_into()
-                .unwrap()
-        }
-
-        #[allow(non_snake_case)]
-        pub fn M() -> Proof {
-            "E4A1C32579947717FAD2306ED9BC7F4307E4FAFA"
-                .try_into()
-                .unwrap()
-        }
-
-        pub fn x() -> PrivateKey {
-            "AE44391953F49241C4FA1FF6E134023397B89905"
-                .try_into()
-                .unwrap()
-        }
-    }
-
-    pub type Mock = Srp6_256Mock;
-
-    #[test]
-    #[cfg(feature = "wow")]
-    fn should_prepare_a_new_user_w_case_insensitive_username() {
-        use crate::dangerous::Srp6_256;
-
-        let (s, v) = Srp6_256::default().generate_new_user_secrets_w_salt(
-            "ADMINISTRATOR",
-            "administrator",
-            Mock::s(),
-        );
-        assert_eq!(s, Mock::s(), "Salt is not correct: {}", s.to_string());
-        assert_eq!(v, Mock::v(), "Verifier is not correct: {}", v.to_string());
-
-        let (s, v) = Srp6_256::default().generate_new_user_secrets_w_salt(
-            "administrator",
-            "administrator",
-            Mock::s(),
-        );
-        assert_eq!(s, Mock::s());
-        assert_eq!(v, Mock::v());
-
-        // with a random salt things are always different
-        let (s, v) =
-            Srp6_256::default().generate_new_user_secrets("administrator", "administrator");
-        assert_ne!(s, Mock::s());
-        assert_ne!(v, Mock::v());
-    }
-
-    #[test]
-    fn should_panic_when_key_length_does_not_fit_to_modulus() {
-        type Srp = Srp6<10, 10>;
-        let err = Srp::new(
-            Generator::from(3),
-            PrimeModulus::from_hex_str_be("FE27").unwrap(),
-        );
-        assert!(err.is_err());
-        assert_eq!(
-            err.err().unwrap(),
-            Srp6Error::KeyLengthMismatch {
-                expected: Srp::KEY_LEN,
-                given: 2
-            }
-        )
-    }
-
-    #[test]
-    #[allow(non_snake_case)]
-    #[cfg(feature = "wow")]
-    fn should_do_the_full_round_trip_to_proof() {
-        use crate::dangerous::Srp6_256;
-
-        // hard mocked
-        let user = mocked_user_details();
-
-        // given
-        let srp6 = Srp6_256::default();
-        let (handshake, proof_verifier) = srp6.start_handshake(&user);
-
-        // when
-        // a client provides proof
-        let user_password: ClearTextPasswordRef = Mock::p();
-        let (proof, strong_proof_verifier) = handshake
-            .calculate_proof(user.username.as_str(), user_password)
-            .unwrap();
-
-        // then
-        // a server verifies users proof
-        let strong_proof = proof_verifier.verify_proof(&proof);
-        assert!(strong_proof.is_ok());
-        let (strong_proof, _key) = strong_proof.unwrap();
-
-        // also the client needs to verify
-        let _session_key = strong_proof_verifier
-            .verify_strong_proof(&strong_proof)
-            .unwrap();
-    }
-
-    #[cfg(feature = "dangerous")]
-    fn mocked_user_details() -> UserSecrets {
-        UserSecrets {
-            username: Mock::I().to_owned(),
-            salt: Mock::s(),
-            verifier: Mock::v(),
-        }
-    }
+    )
 }
